+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Ver 0.0.3
 2015-06-23
 * home page status of parkmeterlebanon corrected
+* added states to zboota notification email, i.e. new ticket, closed ticket, etc.
+ * set sendNtofications to use simulation only so as to watch how this behaves over the next few days
+ * will set simulation to false once I find this to have settled
 
 2015-06-22
 * added api hook to return the web availability ddb table + integrated it into home page